+++ Sources/Swift/Integrations/UserFeedback/SentryUserFeedbackIntegrationDriver.swift
@@ -42,6 +42,16 @@ import UIKit
         } else if let widgetConfigBuilder = configuration.configureWidget {
             widgetConfigBuilder(configuration.widgetConfig)
             validate(configuration.widgetConfig)
+
+            /*
+             * We cannot currently automatically inject a widget into a SwiftUI application, because at the recommended time to start the Sentry SDK (SwiftUIApp.init) there is nowhere to put a UIWindow overlay. SwiftUI apps must currently declare a UIApplicationDelegateAdaptor that returns a UISceneConfiguration, which we can then extract a connected UIScene from into which we can inject a UIWindow.
+             *
+             * At the time this integration is being installed, if there is no UIApplicationDelegate and no connected UIScene, it is very likely we are in a SwiftUI app, but it's possible we could instead be in a UIKit app that has some nonstandard launch procedure or doesn't call SentrySDK.start in a place we expect/recommend, in which case they will need to manually display the widget when they're ready by calling SentrySDK.feedback.showWidget.
+             */
+            if UIApplication.shared.connectedScenes.isEmpty && UIApplication.shared.delegate == nil {
+                return
+            }
+
             if configuration.widgetConfig.autoInject {
                 widget = SentryUserFeedbackWidget(config: configuration, delegate: self)
             }
@@ -57,9 +67,9 @@ import UIKit
     @objc public func showWidget() {
         if widget == nil {
             widget = SentryUserFeedbackWidget(config: configuration, delegate: self)
-        } else {
-            widget?.rootVC.setWidget(visible: true, animated: configuration.animations)
         }
+
+        widget?.rootVC.setWidget(visible: true, animated: configuration.animations)
     }
 
     @objc public func hideWidget() {