+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Fixes
 
 - Set handled to false for fatal app hangs (#5514)
+- User feedback widget automatically injects into SwiftUI apps correctly (#5223)
 
 ### Improvements
 
@@ -145,7 +146,6 @@
 
 - Improved time-to-display tracker to not crash when using view life cycle methods incorrectly (#5048)
 - Enable view renderer V2 by default in session replay and preview redact options when using initializer with default values (#5210)
-- User feedback widget automatically injects into SwiftUI apps correctly (#5223)
 
 ## 8.50.1